Tests: Fix ImageUpload tests after renaming classes.
